Выносит createPreconditions в отдельный файл
