+++ src/mailbox.js
@@ -1,28 +1,7 @@
 const mailboxes = {};
 
 import {createSubscriptions} from './subscriptions';
-
-const createPreconditions = (context, next) => {
-  const preconditions = [];
-
-  return {
-    add(fn) {
-      preconditions.push(fn);
-    },
-    run(message, index = 0) {
-      if (index >= preconditions.length) {
-        next(message);
-      } else {
-        const precondition = preconditions[index];
-        precondition(
-          message,
-          context,
-          () => this.run(message, index + 1)
-        );
-      }
-    }
-  };
-};
+import {createPreconditions} from './preconditions';
 
 const createIdentifiedMailbox = (identity) => {
   const subscriptions = createSubscriptions();
